+++ flake.nix
@@ -46,7 +46,7 @@
     mkSdk = pkgs: {runTests ? false}:
       pkgs.stdenvNoCC.mkDerivation (finalAttrs: {
         pname = "zaptunnel-sdk";
-        version = "0.3.0";
+        version = "0.4.0";
         src = pkgs.lib.cleanSourceWith {
           src = ./sdk;
           filter = path: _type: let